# target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
...
paths:
...
  /accounts/{accountId}:
    delete:
      description: Удаляет учетную запись по ее ID
      operationId: deleteAccount
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          format: int32
          minimum: 1
        style: simple
      responses:
        "400":
          description: "Невалидные параметры запроса, Аккаунт связан с животным"
        "401":
          description: "Неверные авторизационные данные, Запрос от неавторизованно\
            го аккаунта"
        "200":
          description: Учетная запись успешно удалена
        "403":
          description: "Обновление не своего аккаунта, аккаунт не найден"
      summary: Удалить учетную запись
      tags:
      - Accounts
      x-accepts: application/json
      x-tags:
      - tag: Accounts
    get:
      description: Возвращает данные учетной записи по ее ID
      operationId: getAccountId
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          format: int32
          minimum: 1
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Учетная запись успешно найдена
        "401":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Неверные авторизационные данные
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Некорректный ID
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Аккаунт с таким accountId не найден
      summary: Получить учетную запись по ID
      tags:
      - Accounts
      x-accepts: '*/*'
      x-tags:
      - tag: Accounts
    put:
      description: Обновляет данные учетной записи по ее ID
      operationId: updateAccount
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          format: int32
          minimum: 1
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDtoRequest'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Список учетных записей получен
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Невалидные параметры запроса
        "401":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: "Неверные авторизационные данные, Запрос от неавторизованно\
            го аккаунта"
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: Аккаунт с таким email уже существует
        "403":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDtoResponse'
          description: "Обновление не своего аккаунта, аккаунт не найден"
      summary: Обновить учетную запись
      tags:
      - Accounts
      x-content-type: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Accounts
...
  /accounts/search:
    get:
      description: "Позволяет искать учетные записи по имени, фамилии, email с паг\
        инацией"
      operationId: searchAccount
      parameters:
      - explode: true
        in: query
        name: firstName
        required: false
        schema: {}
        style: form
      - explode: true
        in: query
        name: lastName
        required: false
        schema: {}
        style: form
      - explode: true
        in: query
        name: email
        required: false
        schema: {}
        style: form
      - explode: true
        in: query
        name: from
        required: false
        schema:
          default: 0
          format: int32
          minimum: 0
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          format: int32
          minimum: 1
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/AccountDtoResponse'
          description: Список учетных записей получен
        "400":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/AccountDtoResponse'
          description: Невалидные параметры запроса
        "401":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/AccountDtoResponse'
          description: Неверные авторизационные данные
      summary: Поиск учетных записей
      tags:
      - Accounts
      x-accepts: '*/*'
      x-tags:
      - tag: Accounts
...
components:
  schemas:
...
    AccountDtoRequest:
      example:
        firstName: ""
        lastName: ""
        password: ""
        email: ""
      properties:
        firstName: {}
        lastName: {}
        email: {}
        password: {}
      required:
      - email
      - firstName
      - lastName
      - password
    AccountDtoResponse:
      example:
        firstName: ""
        lastName: ""
        id: ""
        email: ""
      properties:
        id:
          format: int32
        firstName: {}
        lastName: {}
        email: {}
...